+++ Backend/Tabibi.Application/DTOs/ChatSessionDetailsDTO.cs
@@ -4,11 +4,11 @@ namespace Tabibi.Application.DTOs
 {
     public class ChatSessionDetailsDTO
     {
-        public int SessionId { get; set; }
+        public long SessionId { get; set; }
         public string DoctorName { get; set; } = string.Empty;
         public string DoctorSpecialty { get; set; } = string.Empty;
         public string PatientName { get; set; } = string.Empty;
-        public int? DoctorId { get; set; }
+        public long? DoctorId { get; set; }
         public string DoctorUserId { get; set; } = string.Empty;
         public string PatientUserId { get; set; } = string.Empty;
         public bool IsCompanyPaid { get; set; }
