+++ Backend/Tabibi.Application/DTOs/ChatSessionSummaryDTO.cs
@@ -4,10 +4,10 @@ namespace Tabibi.Application.DTOs
 {
     public class ChatSessionSummaryDTO
     {
-        public int SessionId { get; set; }
+        public long SessionId { get; set; }
         public string OtherPartyName { get; set; } = string.Empty;
         public string OtherPartyUserId { get; set; } = string.Empty;
-        public int? DoctorId { get; set; }
+        public long? DoctorId { get; set; }
         public string OtherPartySpecialty { get; set; } = string.Empty;
         public string LastMessage { get; set; } = string.Empty;
         public DateTime? LastMessageTime { get; set; }
